backend/requirements/development.txt
-r ./base.txt

django-extensions==3.1.5
djongo==1.3.6

pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.6.1

Faker==9.9.0
ipdb==0.13.9
factory_boy==3.2.1
freezegun==1.1.0


Sphinx==4.3.1
